+++ packages/date-picker/src/DateRangePicker/DateRangePicker.tsx
@@ -1,21 +1,17 @@
 import { Label } from "@kaizen/draft-form"
 import React, { RefObject, useRef, useState } from "react"
 import dateStart from "@kaizen/component-library/icons/date-start.icon.svg"
-import { usePopper } from "react-popper"
 import cx from "classnames"
 import { Icon } from "@kaizen/component-library"
 import { FocusOn } from "react-focus-on"
-import {
-  DateRange,
-  DateInterval,
-  ActiveModifiers,
-  isMatch,
-} from "react-day-picker"
+import { DateRange, DateInterval, isMatch } from "react-day-picker"
+import { enUS } from "date-fns/locale"
 import { calculateDisabledDays } from "../utils/calculateDisabledDays"
 import { isDisabledDate } from "../utils/isDisabledDate"
-import datePickerStyles from "./DatePicker.scss"
-import { Calendar, CalendarProps } from "./components/Calendar"
-import { DayOfWeek } from "./enums"
+import { Calendar, CalendarProps } from "../_primitives/Calendar"
+import { DayOfWeek } from "../enums"
+import { CalendarWrapper } from "../_primitives/CalendarWrapper"
+import dateRangePickerStyles from "./DateRangePicker.scss"
 
 export interface DateRangePickerProps {
   id: string
@@ -24,7 +20,6 @@ export interface DateRangePickerProps {
   isDisabled?: boolean
   buttonRef?: RefObject<HTMLButtonElement>
   description?: string
-
   /** Selected date range which is being updated in handleDayClick and checked
    * if within range/not disabled and then passed back to the client to update
    * the state.
@@ -39,7 +34,7 @@ export interface DateRangePickerProps {
   /** Accepts a DayOfWeek value to start the week on that day. By default,
    * it's set to Monday.
    */
-  weekStartsOn?: DayOfWeek
+  weekStartsOn?: CalendarProps["weekStartsOn"]
 
   // Accepts a date to display that month on first render.
   defaultMonth?: CalendarProps["defaultMonth"]
@@ -85,7 +80,6 @@ export const DateRangePicker: React.VFC<DateRangePickerProps> = ({
   id,
   buttonRef = useRef<HTMLButtonElement>(null),
   description,
-  onChange,
   labelText,
   isDisabled = false,
   classNameOverride,
@@ -99,38 +93,11 @@ export const DateRangePicker: React.VFC<DateRangePickerProps> = ({
   defaultMonth,
   selectedDateRange,
   value,
+  onChange,
   ...inputProps
 }) => {
+  const containerRef = useRef<HTMLInputElement>(null)
   const [isOpen, setIsOpen] = useState(false)
-
-  const [referenceElement, setReferenceElement] =
-    useState<HTMLDivElement | null>(null)
-  const [popperElement, setPopperElement] = useState<HTMLDivElement | null>(
-    null
-  )
-  const wrapperRef = useRef<HTMLDivElement>(null)
-
-  const { styles, attributes } = usePopper(referenceElement, popperElement, {
-    modifiers: [
-      {
-        name: "offset",
-        options: {
-          offset: [0, 15],
-        },
-      },
-    ],
-    placement: "bottom-start",
-  })
-
-  const handleOpenClose = () => {
-    setIsOpen(!isOpen)
-  }
-
-  const handleReturnFocus = () => {
-    if (buttonRef.current) {
-      buttonRef.current.focus()
-    }
-  }
 
   const disabledDays = calculateDisabledDays({
     disabledDates,
@@ -141,7 +108,17 @@ export const DateRangePicker: React.VFC<DateRangePickerProps> = ({
     disabledAfter,
   })
 
-  const handleDayClick = (day: Date, modifiers: ActiveModifiers) => {
+  const handleOpenClose = (): void => {
+    setIsOpen(!isOpen)
+  }
+
+  const handleReturnFocus = (): void => {
+    if (buttonRef.current) {
+      buttonRef.current.focus()
+    }
+  }
+
+  const handleDayClick = (day: Date): void => {
     /** react-day-picker will fire events for disabled days by default.
      *  We're checking here if it includes the CSS Modules class for disabled
      *  on the modifier to then return early.
@@ -171,7 +148,7 @@ export const DateRangePicker: React.VFC<DateRangePickerProps> = ({
     }
   }
 
-  const isSelectingFirstDay = (range: DateRange, day: Date) => {
+  const isSelectingFirstDay = (range: DateRange, day: Date): boolean => {
     const isBeforeFirstDay =
       !!range.from &&
       isMatch(day, [
@@ -191,34 +168,35 @@ export const DateRangePicker: React.VFC<DateRangePickerProps> = ({
   }
 
   return (
-    <div ref={wrapperRef}>
-      <div ref={setReferenceElement}>
+    <div>
+      <div ref={containerRef}>
         <Label disabled={isDisabled} htmlFor={id} labelText={labelText} />
         <button
           id={id}
           className={cx(
-            datePickerStyles.button,
-            datePickerStyles.withStartIconAdornment,
+            dateRangePickerStyles.button,
+            dateRangePickerStyles.withStartIconAdornment,
             {
-              [datePickerStyles.disabled]: isDisabled,
+              [dateRangePickerStyles.disabled]: isDisabled,
             }
           )}
           disabled={isDisabled}
           ref={buttonRef}
           onClick={handleOpenClose}
-          {...inputProps}
           aria-label={
             selectedDateRange?.from ? `Change date: ${value}` : "Choose date"
           }
+          {...inputProps}
         >
-          <div className={datePickerStyles.startIconAdornment}>
+          <div className={dateRangePickerStyles.startIconAdornment}>
             <Icon icon={dateStart} role="presentation" />
           </div>
-          <span className={datePickerStyles.value}>
+          <span className={dateRangePickerStyles.value}>
             {selectedDateRange?.from ? value : undefined}
           </span>
         </button>
       </div>
+
       {isOpen && (
         <FocusOn
           scrollLock={false}
@@ -229,21 +207,21 @@ export const DateRangePicker: React.VFC<DateRangePickerProps> = ({
           onEscapeKey={() => {
             handleOpenClose()
           }}
+          enabled={isOpen}
         >
-          <Calendar
-            mode="range"
-            id="calendar-dialog"
-            setPopperElement={setPopperElement}
-            popperStyles={styles}
-            popperAttributes={attributes}
-            classNameOverride={classNameOverride}
-            defaultMonth={defaultMonth}
-            weekStartsOn={weekStartsOn}
-            disabledDays={disabledDays}
-            modifiers={modifiers}
-            selectedRange={selectedDateRange}
-            onDayChange={handleDayClick}
-          />
+          <CalendarWrapper referenceElement={containerRef.current}>
+            <Calendar
+              mode="range"
+              id={`${id}-calendar-dialog`}
+              selectedRange={selectedDateRange}
+              defaultMonth={defaultMonth}
+              weekStartsOn={weekStartsOn}
+              disabledDays={disabledDays}
+              onDayChange={handleDayClick}
+              modifiers={modifiers}
+              locale={enUS}
+            />
+          </CalendarWrapper>
         </FocusOn>
       )}
     </div>